Update templating to pin to only major version and use a single template
It's extremely unlikely that memcached is going to maintain more than one major version at a time (even that seems to be a bit much sometimes, which is relatable).

This also converts the Dockerfiles to use semicolons instead of `&&` chains (see https://github.com/docker-library/memcached/compare/b8a7264..00003b7 to compare this without that conversion).

Additionally, this converts us to just two variants and auto-detects the latest Alpine/Debian, since we don't have a good reason to support more than one of each Debian and Alpine.